Updated from Welcome. Streamlined from previous version and worked out some quirks while I was doing manual testing. Slight addition of documentation.
