+++ src/Task.java
@@ -3,12 +3,15 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 
 /**
- * Collation of all information for a Task from User.
+ * This Task class models object task with all its component.
+ * It provides a platform to provide the mould of how a task should
+ * be created and access to each of the components in a task for
+ * any update necessary.
  *
  * @author Zhi Chen
- * @version 2018.09.22
+ * @2018.10.12
  */
-public class Task implements Comparable<Task>{
+public class Task implements Comparable<Task> {
 
 
     private String title;
@@ -19,40 +22,44 @@ public class Task implements Comparable<Task>{
 
     /**
      * Collects information needed for Task.
-     * @param title Summary of required task.
-     * @param dueDate Deadline of Task.
+     *
+     * @param title       Summary of required task.
+     * @param dueDate     Deadline of Task.
      * @param description Detailed information of Task.
      */
     public Task(String title, String dueDate,
-                String description, String projectTitle){
+                String description, String projectTitle) {
 
         this.title = title;
         this.dueDate = createDate(dueDate);
         this.description = description;
-        this.projectTitle =projectTitle;
+        this.projectTitle = projectTitle;
         this.done = false;
     }
 
     /**
      * Accessor method for title of task.
+     *
      * @return String title.
      */
-    public String getTitle(){
+    public String getTitle() {
 
         return title;
     }
 
     /**
      * Mutator method to set new task title.
+     *
      * @param newTitle User entry for new task title
      */
-    public void setTitle(String newTitle){
+    public void setTitle(String newTitle) {
 
         this.title = newTitle;
     }
 
     /**
      * Conversion of String dueDate to Date type.
+     *
      * @param dueDate User entry in String format: dd-MMM-yyyy
      * @return dueDate of type Date.
      */
@@ -60,47 +67,49 @@ public class Task implements Comparable<Task>{
 
         SimpleDateFormat formatter = new SimpleDateFormat("dd-MMM-yyyy");
 
-        try
-        {
+        try {
             return formatter.parse(dueDate);
-        }
-        catch (ParseException e)
-        {
-            e.printStackTrace();
+        } catch (ParseException e) {
+            System.err.println("Invalid Deadline. Please update with " +
+                    "valid date format under Edit function.");
             return null;
         }
     }
 
     /**
      * Accessor for deadline.
+     *
      * @return dueDate of Date type.
      */
-    public Date getDueDate(){
+    public Date getDueDate() {
 
         return dueDate;
     }
 
     /**
      * Mutator method to set new deadline.
+     *
      * @param newDueDate Collects String dueDate from user and converts
      *                   its type to Date.
      */
-    public void setDueDate(String newDueDate){
+    public void setDueDate(String newDueDate) {
 
         this.dueDate = createDate(newDueDate);
     }
 
     /**
      * Acessor for details of Task.
+     *
      * @return Additional information of task.
      */
-    public  String getDescription(){
+    public String getDescription() {
 
         return description;
     }
 
     /**
      * Mutator method to edit additional information of task.
+     *
      * @param newDescription Edition to task description.
      */
     public void setDescription(String newDescription) {
@@ -110,28 +119,31 @@ public class Task implements Comparable<Task>{
 
     /**
      * Accessor method for Project Title.
+     *
      * @return The project task is currently under.
      */
-    public String getProjectTitle(){
+    public String getProjectTitle() {
 
         return projectTitle;
     }
 
     /**
      * Mutator method to change task to another project.
+     *
      * @param newProjectTitle Project that task is shifted to.
      */
-    public void setProjectTitle(String newProjectTitle){
+    public void setProjectTitle(String newProjectTitle) {
 
         this.projectTitle = newProjectTitle;
     }
 
     /**
      * Accessor method with print line to indicate if task is completed.
+     *
      * @return Verification of task status.
      */
-    public String getStatus(){
-        if (!done){
+    public String getStatus() {
+        if (!done) {
             return "Pending action.";
         } else {
             return "Completed.";
@@ -140,36 +152,40 @@ public class Task implements Comparable<Task>{
 
     /**
      * Accessor method to get boolean status of task.
+     *
      * @return The status of task.
      */
-    public boolean doneStatus(){
+    public boolean doneStatus() {
 
         return done;
     }
 
     /**
      * Mutator method to change status of task to completed.
+     *
      * @return Task completed.
      */
-    public boolean taskDone(){
+    public boolean taskDone() {
 
         return done = true;
     }
 
     /**
      * Mutator method to change status of task to pending.
+     *
      * @return Task pending.
      */
-    public boolean taskReset(){
+    public boolean taskReset() {
 
         return done = false;
     }
 
     /**
      * Accesor method to get String version of dueDate
+     *
      * @return String format date
      */
-    public String getDateToPrint(){
+    public String getDateToPrint() {
 
         SimpleDateFormat formatter = new SimpleDateFormat("dd-MMM-yyyy");
         return formatter.format(dueDate);
@@ -177,14 +193,15 @@ public class Task implements Comparable<Task>{
 
     /**
      * Prints entire task details.
+     *
      * @return Entire task details.
      */
     @Override
-    public String toString(){
+    public String toString() {
 
         String dateToPrint = getDateToPrint();
         String task = "Project Title: " + projectTitle + "\n" +
-                "Task title: "+ title + "\n" + "Deadline: "+ dateToPrint + "\n" +
+                "Task title: " + title + "\n" + "Deadline: " + dateToPrint + "\n" +
                 "Description: " + description + "\n" + "Status: " + getStatus() + "\n";
 
         return task;
@@ -192,11 +209,12 @@ public class Task implements Comparable<Task>{
 
     /**
      * Compares dueDate in task.
+     *
      * @param task Task details input by user.
      * @return integer value used to compare the dueDates.
      */
     @Override
-    public int compareTo(Task task){
+    public int compareTo(Task task) {
 
         return this.dueDate.compareTo(task.getDueDate());
     }
